+++ changelog.md
@@ -4,3 +4,5 @@
 * Discord reports are now supported, how fun (not accessable yet)!
 * removed a lot of the more "fun" features. (like /when day breaks or stuff like that, maybe ill bring it back)
 * I haven't implemented a lot of the old features from v1, but I will eventually.
+* lol forgot expiring
+* console??? added.